move in place assignment functions back to AssignmentGen
